fix(ci): ruff lint/format + i18n parity + JS test glob
CI failures on #416:
- ruff 0.16.8: RUF046 redundant int(round(...)) in trace_observer,
  inference, decision_executor (round() already int in py3); RUF100
  unused noqa in the SSE harness; ruff-format whitespace in 3 files
- js-tests.yml globs tests/js/*.test.js — renamed the trace graph
  suite from .test.mjs so CI picks it up (27/27 pass under that glob,
  incl. the 25 trace tests)
- frontend_i18n_parity: nav.feature.decision-trace missing fa/de/es/ar
  — added to all four locale scopes

Verified: ruff check/format clean repo-wide, tsc 0 errors, vite build
ok, all tests/js/*.test.js pass, pytest 9/9.

+++ src/nexus_scalp/application/live/inference.py
@@ -415,7 +415,7 @@ class InferenceService:
                 component="inference",
                 event_type="MODEL_INFERENCE",
                 status="OK",
-                latency_us=int(round(float(self._last_e2e_ms) * 1000.0)),
+                latency_us=round(float(self._last_e2e_ms) * 1000.0),
                 detail=_d,
             )
         return probs
